+++ src/ToledoMessage/Models/ConversationReadPointer.cs
@@ -2,8 +2,8 @@ namespace ToledoMessage.Models;
 
 public class ConversationReadPointer
 {
-    public decimal UserId { get; set; }
-    public decimal ConversationId { get; set; }
+    public long UserId { get; set; }
+    public long ConversationId { get; set; }
     public long LastReadSequenceNumber { get; set; }
     public int UnreadCount { get; set; }
     public DateTimeOffset? LastReadAt { get; set; }
